Move HTTP handler functions into 2 new files
